2 files modified, 5 added,locate components to separate files
