+++ src/App.js
@@ -1,18 +1,16 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom';
-import {
-  FaSearch,
-  FaBars,
-  FaUserAlt,
-  FaLink,
-  FaMoneyBillAlt
-} from 'react-icons/fa';
+import { FaSearch, FaBars, FaUserAlt } from 'react-icons/fa';
 import { Wave } from 'react-animated-text';
 import './style.css';
 import Products from './components/Products';
 import Contact from './components/Contact.js';
 import Delivery from './components/Delivery.js';
 import About from './components/About.js';
+import Main from './components/Main.js';
+import Menu from './components/Menu.js';
+import SideNav from './components/Sidenav.js';
+import Admin from './components/Admin.js';
 import productsArray from './components/constants.js';
 
 const App = () => {
@@ -178,178 +176,3 @@ const App = () => {
   );
 };
 export default App;
-
-export const Admin = props => {
-  const [imgUrl, setImgUrl] = useState('');
-  const [price, setPrice] = useState('');
-  const [title, setTitle] = useState('');
-
-  function handleChange(e) {
-    let val = e.currentTarget.value;
-    if (e.currentTarget.className == 'url') setImgUrl(val);
-    else if (e.currentTarget.className == 'title') setTitle(val);
-    else setPrice(val);
-  }
-  return (
-    <div className="common admin">
-      <form
-        onSubmit={e => {
-          e.preventDefault();
-          props.addProduct([imgUrl, title, price]);
-        }}
-      >
-        <label>პროდუქტის დამატება</label>
-        <div>
-          {<FaLink />}
-          <input
-            className="url"
-            onChange={handleChange}
-            value={imgUrl}
-            placeholder="Image url"
-          />
-        </div>
-        <div>
-          {<FaLink />}
-          <input
-            className="title"
-            onChange={handleChange}
-            value={title}
-            placeholder="Title"
-          />
-        </div>
-        <div>
-          {<FaMoneyBillAlt />}
-          <input
-            className="price"
-            onChange={handleChange}
-            value={price}
-            placeholder="Price"
-          />
-        </div>
-        <button>დამატება</button>
-      </form>
-      <div id="adminDelText">პროდუქტის წაშლა ან ფასის შეცვლა</div>
-      <Products arr={[...props.arr]} />
-    </div>
-  );
-};
-
-export const Main = () => {
-  return (
-    <div className="common">
-      <img src="https://www.coronainsights.com/wp-content/uploads/2014/11/Welcome.jpg" />
-    </div>
-  );
-};
-
-export const SideNav = props => {
-  const setSidenavWidth = props.arr[0];
-  const styles = {
-    height: props.arr[3]
-  };
-
-  return (
-    <div className="sidenav" style={styles}>
-      <div id="logo" />
-      <a className="closebtn" onClick={props.arr[4]}>
-        <div id="closeIcon">&times;</div>
-      </a>
-      {/**This component is used two times with different arguments in header,here and below */}
-      <Menu arr={[...props.arr]} />
-      <Link
-        to="/admin"
-        onClick={() => {
-          setSidenavWidth(0);
-        }}
-      >
-        <div>
-          <span>ადმინის პანელი</span>
-        </div>
-      </Link>
-    </div>
-  );
-};
-
-const Menu = props => {
-  let setSidenavWidth = props.arr[0],
-    activeMenuOption = props.arr[1],
-    setActiveMenuOption = props.arr[2];
-  const Border = 'border-bottom';
-  const Style = {
-    /*'borderBottom': '3px solid rgb(35, 167, 75)'*/
-    background: 'rgb(12, 136, 185)',
-    color: 'white'
-  };
-  return (
-    <>
-      <Link
-        to="/"
-        style={activeMenuOption == 0 ? Style : {}}
-        onClick={() => {
-          {
-            setActiveMenuOption(0);
-          }
-          {
-            setSidenavWidth(0);
-          }
-        }}
-      >
-        <div className="options">მთავარი</div>
-      </Link>
-      <Link
-        to="/products"
-        style={activeMenuOption == 1 ? Style : {}}
-        onClick={() => {
-          {
-            setActiveMenuOption(1);
-          }
-          {
-            setSidenavWidth(0);
-          }
-        }}
-      >
-        <div className="options">პროდუქცია</div>
-      </Link>
-      <Link
-        to="/delivery"
-        style={activeMenuOption == 2 ? Style : {}}
-        onClick={() => {
-          {
-            setActiveMenuOption(2);
-          }
-          {
-            setSidenavWidth(0);
-          }
-        }}
-      >
-        <div className="options">მიწოდების სერვისი</div>
-      </Link>
-      <Link
-        to="/about"
-        style={activeMenuOption == 3 ? Style : {}}
-        onClick={() => {
-          {
-            setActiveMenuOption(3);
-          }
-          {
-            setSidenavWidth(0);
-          }
-        }}
-      >
-        <div className="options">ჩვენ შესახებ</div>
-      </Link>
-      <Link
-        to="/contact"
-        style={activeMenuOption == 4 ? Style : {}}
-        onClick={() => {
-          setActiveMenuOption(4);
-          {
-            setSidenavWidth(0);
-          }
-        }}
-      >
-        <div className="options">საკონტაქტო ინფორმაცია</div>
-      </Link>
-    </>
-  );
-};
