+++ src/components/Main.js
@@ -1,0 +1,9 @@
+import React from 'react';
+const Main = () => {
+  return (
+    <div className="common">
+      <img src="https://www.coronainsights.com/wp-content/uploads/2014/11/Welcome.jpg" />
+    </div>
+  );
+};
+export default Main;
