+++ src/components/Products.js
@@ -1,18 +1,6 @@
 import React, { useState, useEffect } from 'react';
+import MySelect from './MySelect';
 import '../style.css';
-
-const MySelect = props => {
-  return (
-    <select
-      onClick={e => props.onClick(e.currentTarget.value)}
-      className="select"
-    >
-      {props.options.map((e, i) => {
-        return <option value={e}>{e}</option>;
-      })}
-    </select>
-  );
-};
 
 const Products = props => {
   const [newPrice, setNewPrice] = useState('');
